backend/pyproject.toml
[project]
dependencies = [
  "fastapi[standard]",
  "sqlalchemy",
  "pydantic",
  "pydantic-settings",
  "psycopg2-binary",
  "alembic",
  "httpx",
  "passlib[bcrypt]",
  "PyJWT",
]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"

[project.optional-dependencies]
dev = ["ruff", "basedpyright"]

# ============================================================
# pytest
# ============================================================
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-ra --tb=short"

# ============================================================
# Ruff (Linter + Formatter)
# ============================================================
[tool.ruff]
target-version = "py313"
line-length = 88
# Lint設定
[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "W",   # pycodestyle warnings
  "F",   # Pyflakes
  "I",   # isort
  "B",   # flake8-bugbear
  "C4",  # flake8-comprehensions
  "UP",  # pyupgrade
  "SIM", # flake8-simplify
]
ignore = [
  "E501", # line too long
]
# isort設定
[tool.ruff.lint.isort]
known-first-party = ["app"]
# Format設定
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
# ============================================================
# basedpyright (Type Checker)
# ============================================================
[tool.basedpyright]
pythonVersion = "3.13"
venvPath = "."
venv = ".venv"
stubPath = "typings"
typeCheckingMode = "standard"
reportMissingImports = true
reportMissingTypeStubs = false
# 型チェックの対象/除外
include = ["app", "tests"]
exclude = ["**/__pycache__", "**/.venv"]
# 厳密さの調整（必要に応じて）
# reportUnknownMemberType = "warning"
# reportUnknownArgumentType = "warning"
